# antimoodlo-api/docs/swagger.yaml
definitions:
...
  Models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  Models.Question:
    properties:
      id:
        type: integer
      questiontext:
        type: string
      questiontypeid:
        type: integer
      quizid:
        type: integer
    type: object
...
  Models.Quiz:
    properties:
      courseid:
        type: integer
      duration:
        type: integer
      enddate:
        type: string
      id:
        type: integer
      maxgrade:
        type: integer
      startdate:
        type: string
      stateid:
        type: integer
      submiteddate:
        type: string
      title:
        type: string
    type: object
  Models.User:
    properties:
      userid:
        type: integer
      userlogin:
        type: string
      userpassword:
        type: string
      userrole:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /questions/{id}:
    delete:
      parameters:
      - description: ID вопроса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Удалить вопрос
      tags:
      - Questions
    put:
      consumes:
      - application/json
      parameters:
      - description: ID вопроса
        in: path
        name: id
        required: true
        type: integer
      - description: Вопрос
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/Models.Question'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.Question'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Обновить вопрос
      tags:
      - Questions
...
  /quizzes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Models.Quiz'
            type: array
      summary: Получить все тесты
      tags:
      - Quizzes
    post:
      consumes:
      - application/json
      parameters:
      - description: Тест
        in: body
        name: quiz
        required: true
        schema:
          $ref: '#/definitions/Models.Quiz'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Models.Quiz'
      summary: Создать тест
      tags:
      - Quizzes
  /quizzes/{id}:
    delete:
      parameters:
      - description: ID теста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Удалить тест
      tags:
      - Quizzes
    get:
      parameters:
      - description: ID теста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.Quiz'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Получить тест по ID
      tags:
      - Quizzes
    put:
      consumes:
      - application/json
      parameters:
      - description: ID теста
        in: path
        name: id
        required: true
        type: integer
      - description: Тест
        in: body
        name: quiz
        required: true
        schema:
          $ref: '#/definitions/Models.Quiz'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.Quiz'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Обновить тест
      tags:
      - Quizzes
  /quizzes/{id}/questions:
    get:
      parameters:
      - description: ID теста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Models.Question'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Получить вопросы теста
      tags:
      - Questions
    post:
      consumes:
      - application/json
      parameters:
      - description: ID теста
        in: path
        name: id
        required: true
        type: integer
      - description: Вопрос
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/Models.Question'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Models.Question'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Добавить вопрос в тест
      tags:
      - Questions
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Models.User'
            type: array
      summary: Получить всех пользователей
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: Пользователь
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/Models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Models.User'
      summary: Создать нового пользователя
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Удалить пользователя
      tags:
      - Users
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Получить пользователя по ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Пользователь
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/Models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Обновить пользователя
      tags:
      - Users
  /users/login/{login}:
    get:
      parameters:
      - description: Логин пользователя
        in: path
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Models.ErrorResponse'
      summary: Получить пользователя по логину
      tags:
      - Users
swagger: "2.0"